+++ Packages/Syntax/Package.swift
@@ -13,16 +13,25 @@ let package = Package(
     ],
     dependencies: [
         .package(name: "EditorCore", path: "../EditorCore"),
+        
         .package(url: "https://github.com/apple/swift-argument-parser", from: Version(1, 7, 0)),
         .package(url: "https://github.com/jpsim/Yams", from: Version(6, 2, 0)),
         .package(url: "https://github.com/SimplyDanny/SwiftLintPlugins", from: Version(0, 62, 0)),
+        .package(url: "https://github.com/tree-sitter/swift-tree-sitter", from: Version(0, 9, 0)),
+        
+        .package(url: "https://github.com/alex-pinkus/tree-sitter-swift", branch: "with-generated-files"),
     ],
     targets: [
         .target(
             name: "Syntax",
-            dependencies: ["EditorCore", "Yams"],
-            resources: [.process("Resources")]),
-        .testTarget(name: "SyntaxTests", dependencies: ["Syntax"], resources: [.copy("Syntaxes")]),
+            dependencies: [
+                "EditorCore",
+                "Yams",
+                .product(name: "SwiftTreeSitterLayer", package: "swift-tree-sitter"),
+                .product(name: "TreeSitterSwift", package: "tree-sitter-swift"),
+            ],
+            resources: [.process("Resources"), .copy("Syntaxes")]),
+        .testTarget(name: "SyntaxTests", dependencies: ["Syntax", "EditorCore"], resources: [.copy("Syntaxes")]),
         
         .executableTarget(
             name: "SyntaxMapBuilder",